升级AndroidX

+++ http/src/main/java/cn/oahttp/callback/FileCallBack.java
@@ -3,8 +3,6 @@ package cn.oahttp.callback;
 import android.os.Environment;
 import android.util.Log;
 
-import cn.oahttp.HandlerQueue;
-
 import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
@@ -13,9 +11,10 @@ import java.io.UnsupportedEncodingException;
 import java.net.SocketException;
 import java.net.URLDecoder;
 
+import cn.oahttp.ClientFactory;
+import cn.oahttp.HandlerQueue;
 import okhttp3.Call;
 import okhttp3.Callback;
-import okhttp3.HttpUrl;
 import okhttp3.Response;
 
 /**
@@ -30,7 +29,7 @@ public abstract class FileCallBack implements Callback {
 
     String fileName = null;
 
-    String saveDir = Environment.getExternalStorageDirectory().getAbsolutePath();
+    public static String saveDir = Environment.getExternalStorageDirectory().getAbsolutePath() + "/" + ClientFactory.getApplication().getPackageName();
 
     public FileCallBack() {
     }
@@ -58,59 +57,71 @@ public abstract class FileCallBack implements Callback {
 
     @Override
     public void onResponse(final Call call, final Response response) throws IOException {
+        Log.i("logtag", "============onResponse===========" + response.headers());
         try {
             if (response.isSuccessful()) {
                 //获取文件名称
                 if (fileName == null) {
-                    fileName = getFileName(response, call.request().url());
+                    fileName = getFileName(response, call.request().url().toString());
                 }
                 //文件长度
                 final long totalSize = response.body().contentLength();
                 InputStream is = null;
-                byte[] buf = new byte[2048];
+                byte[] buf = new byte[1024];
                 int len = 0;
                 FileOutputStream fos = null;
                 is = response.body().byteStream();
                 Log.i("logtag", "byteStream======" + fileName);
                 final File file = new File(saveDir, fileName);
-                fos = new FileOutputStream(file);
-                long sum = 0;
-                while ((len = is.read(buf)) != -1) {
-                    sum += len;
-                    fos.write(buf, 0, len);
-                    final long currentSize = sum;
+                if(file.getParent()==null){
+                Log.i("logtag", "byteStream======" + (file.length() == totalSize));
+                new File(saveDir).mkdirs();
+                }
+                if (file.exists() && file.length() == totalSize) {
+                    is.close();
+                    onSuccess(file);
+                } else {
+                    file.createNewFile();
+                    fos = new FileOutputStream(file);
+                    long sum = 0;
+                    while ((len = is.read(buf)) != -1) {
+                        //sum += len;
+                        fos.write(buf, 0, len);
+                        //final long currentSize = sum;
+                        //HandlerQueue.onResultCallBack(new Runnable() {
+                        //    @Override
+                        //    public void run() {
+                        //        double percent = (currentSize * 1.0) / (1.0 * totalSize) * 100;
+                        //        onProgress(totalSize, currentSize, percent);
+                        //    }
+                        //});
+                    }
+                    //fos.flush();
+                    //is.close();
                     HandlerQueue.onResultCallBack(new Runnable() {
                         @Override
                         public void run() {
-                            double percent = (currentSize * 1.0) / (1.0 * totalSize) * 100;
-                            onProgress(totalSize, currentSize, percent);
+                    onSuccess(file);
                         }
                     });
+
                 }
-                fos.flush();
-                is.close();
-                HandlerQueue.onResultCallBack(new Runnable() {
-                    @Override
-                    public void run() {
-                        onSuccess(file);
-                    }
-                });
             } else {
-                HandlerQueue.onResultCallBack(new Runnable() {
-                    @Override
-                    public void run() {
+                //HandlerQueue.onResultCallBack(new Runnable() {
+                //    @Override
+                //    public void run() {
                         onFail(new Exception("网络请求失败！错误代码：" + response.code() + " 信息：" + response.message()));
-                    }
-                });
+                //    }
+                //});
             }
         } catch (SocketException e) {
             e.printStackTrace();
-            HandlerQueue.onResultCallBack(new Runnable() {
-                @Override
-                public void run() {
+            //HandlerQueue.onResultCallBack(new Runnable() {
+            //    @Override
+            //    public void run() {
                     onFail(new Exception("下载请求被取消：tag : " + call.request().tag()));
-                }
-            });
+                //}
+            //});
         }
 
 
@@ -123,8 +134,8 @@ public abstract class FileCallBack implements Callback {
      * @param url
      * @return
      */
-    public String getFileName(Response response, HttpUrl url) {
-        fileName = System.currentTimeMillis() + "";
+    public static String getFileName(Response response, String url) {
+        String fileName = System.currentTimeMillis() + "";
         String cd = response.header("Content-Disposition");
         if (cd != null) {
             String[] split = cd.split(";");
@@ -135,6 +146,9 @@ public abstract class FileCallBack implements Callback {
                     fileName = decodeForResponse(response, fileName);
                 }
             }
+        } else if (setNameIfInUrl(url) != null) {
+            fileName = setNameIfInUrl(url);
+            fileName = decodeForResponse(response, fileName);
         } else if (setNameIfInUrl(url) != null) {
             fileName = setNameIfInUrl(url);
             fileName = decodeForResponse(response, fileName);
@@ -150,7 +164,7 @@ public abstract class FileCallBack implements Callback {
      * @param url
      * @return
      */
-    public String decodeForResponse(Response response, String url) {
+    public static String decodeForResponse(Response response, String url) {
         String encode = "UTF-8";
         if (response.body().contentType().charset() != null) {
             //检查头部contentType有没有编码信息，没有则默认为UTF-8
@@ -171,8 +185,8 @@ public abstract class FileCallBack implements Callback {
      *
      * @param url
      */
-    private String setNameIfInUrl(HttpUrl url) {
-        String s = url.toString();
+    private static String setNameIfInUrl(String url) {
+        String s = url;
         s = s.substring(s.lastIndexOf("/") + 1, s.length());
         if (s.contains("?")) {
             s = s.substring(0, s.lastIndexOf("?"));
